requirements.txt
selenium
requests
ytmusicapi
Pillow
eyed3
colorama
moviepy
yt-dlp
yandex-music
webdriver-manager